restructure alerts popup selector and fix performance issues
